Update README.md
# cmsc11-srs
This program was made for our Final project in CMSC11. This is our own made Student Resgitration Program where you can create and enroll students to a subject. With the help of this program, we can now easily finish our task specially teachers nowadays are busy and there are much things to be done.With this, in just a click click in your keyboard you can directly convert in into csv file.
This program can Create or Update Student Record- You can add and update the name, surname, section, and 3 of his/her scores. More importanly the student ID is already randomize by the program and is surely unique.
This Program can also Search student Record - By just inputing the Student surname,you can directly search for that specific student. Take note that this program is case insensitive because we are ready for any users behavior.
This Program can also Delete Student Record - By just inputing the Student ID which is diplayed in the console you can easly choose which student to delete
This program can also Create and Update Subject Record - You can add or update the Subject Code and Subject Title. More importanly the subject ID is already randomize by the program and is surely unique.Take note that all the student are already enrolled to CMSC11 and you can only add two more subject to the student.
This Program can also Search Subject Record - By just inputing the Subject Code,you can directly search for that specific subject. Take note that this program is case insensitive because we are ready for any users behavior.
This Program can also Delete SUbject Record - By just inputing the Subject ID which is diplayed in the console you can easly choose which student to delete. Take note that CMCS11 can never be deleted beacuse this is a prerequisite to every student.
This Program can also Display All Student and Subject Records
This Program can enroll or drop student to a subject and directly update the changes to the file itself.
And lasty, All the things that you have made can be imported and exported to a files and convert it into csv.
I hope you can learn with the help of this program :):)